refactor(section-level): rename sectionlevel2 to sectionlevel
Jira: XDS-144
