+++ lib/documentLevel.js
@@ -3,7 +3,7 @@
 var headerLevel = require('./headerLevel');
 var fieldLevel = require('./fieldLevel');
 var leafLevel = require('./leafLevel');
-var sectionLevel2 = require('./sectionLevel22');
+var sectionLevel = require('./sectionLevel');
 var contentModifier = require('./contentModifier');
 var condition = require('./condition');
 
@@ -90,61 +90,61 @@ exports.ccd = function(html_renderer) {
           key: 'structuredBody',
           content: [
             [
-              sectionLevel2.allergiesSectionEntriesRequired(
+              sectionLevel.allergiesSectionEntriesRequired(
                 html_renderer.allergiesSectionEntriesRequiredHtmlHeader,
                 html_renderer.allergiesSectionEntriesRequiredHtmlHeaderNA
               ),
               required
             ],
             [
-              sectionLevel2.medicationsSectionEntriesRequired(
+              sectionLevel.medicationsSectionEntriesRequired(
                 html_renderer.medicationsSectionEntriesRequiredHtmlHeader,
                 html_renderer.medicationsSectionEntriesRequiredHtmlHeaderNA
               ),
               required
             ],
             [
-              sectionLevel2.problemsSectionEntriesRequired(
+              sectionLevel.problemsSectionEntriesRequired(
                 html_renderer.problemsSectionEntriesRequiredHtmlHeader,
                 html_renderer.problemsSectionEntriesRequiredHtmlHeaderNA
               ),
               required
             ],
             [
-              sectionLevel2.proceduresSectionEntriesRequired(
+              sectionLevel.proceduresSectionEntriesRequired(
                 html_renderer.proceduresSectionEntriesRequiredHtmlHeader,
                 html_renderer.proceduresSectionEntriesRequiredHtmlHeaderNA
               ),
               required
             ],
             [
-              sectionLevel2.resultsSectionEntriesRequired(
+              sectionLevel.resultsSectionEntriesRequired(
                 html_renderer.resultsSectionEntriesRequiredHtmlHeader,
                 html_renderer.resultsSectionEntriesRequiredHtmlHeaderNA
               ),
               required
             ],
-            sectionLevel2.encountersSectionEntriesOptional(
+            sectionLevel.encountersSectionEntriesOptional(
               html_renderer.encountersSectionEntriesOptionalHtmlHeader,
               html_renderer.encountersSectionEntriesOptionalHtmlHeaderNA
             ),
-            sectionLevel2.immunizationsSectionEntriesOptional(
+            sectionLevel.immunizationsSectionEntriesOptional(
               html_renderer.immunizationsSectionEntriesOptionalHtmlHeader,
               html_renderer.immunizationsSectionEntriesOptionalHtmlHeaderNA
             ),
-            sectionLevel2.payersSection(
+            sectionLevel.payersSection(
               html_renderer.payersSectionHtmlHeader,
               html_renderer.payersSectionHtmlHeaderNA
             ),
-            sectionLevel2.planOfCareSection(
+            sectionLevel.planOfCareSection(
               html_renderer.planOfCareSectionHtmlHeader,
               html_renderer.planOfCareSectionHtmlHeaderNA
             ),
-            sectionLevel2.socialHistorySection(
+            sectionLevel.socialHistorySection(
               html_renderer.socialHistorySectionHtmlHeader,
               html_renderer.socialHistorySectionHtmlHeaderNA
             ),
-            sectionLevel2.vitalSignsSectionEntriesOptional(
+            sectionLevel.vitalSignsSectionEntriesOptional(
               html_renderer.vitalSignsSectionEntriesOptionalHtmlHeader,
               html_renderer.vitalSignsSectionEntriesOptionalHtmlHeaderNA
             )
